[Flash-664] Enable DDL for engine DeltaTree (#526)
- [x] Enable unittest in gtest_dbms
- [x] Enable mock test in tests/delta-merge-test
- [x] Enable fullstack-test/ddl
- [x] Enable fullstack-test/inject (Imported in https://github.com/pingcap/tics/pull/443)
- [x] Refine exception while read / write to DeltaTree (FLASH-994)
  * Use `Exception::addMessage` to add more diagnostics for locate which table is wrong (commit: 716ae4a)
- [x] shutdown should cancel all background tasks (FLASH-995) (commit: 7470c2f)
- [x] Run schrodinger/sddl test

## Others 
* Add atomic-rename table test in `tests/fullstack-test/fault-inject/rename-table.test`, but did not enable this. We will fix it later.
* "dt" engine ONLY support disable_bg_flush = true.
If background flush is enabled, read will not triggle schema sync. Which means that we may not get the right result with out-dated schema.
* Found that 'zero' value of type year is not match with tikv (FLASH-1023)

+++ dbms/src/Storages/Transaction/RegionTable.cpp
@@ -244,7 +244,10 @@ void RegionTable::removeRegion(const RegionID region_id)
         TableID table_id = it->second;
         auto & table = tables.find(table_id)->second;
 
+        do
         {
+            /// Some region of this table is removed, if it is a DeltaTree, write deleteRange.
+
             /// Now we assume that StorageDeltaMerge::deleteRange do not block for long time and do it in sync mode.
             /// If this block for long time, consider to do this in background threads.
             TMTContext & tmt = context->getTMTContext();
@@ -253,22 +256,28 @@ void RegionTable::removeRegion(const RegionID region_id)
             {
                 // acquire lock so that no other threads can change storage's structure
                 auto storage_lock = storage->lockStructure(true, __PRETTY_FUNCTION__);
+                // Check if it has been dropped by other thread
+                if (storage->is_dropped)
+                    break; // continue to remove region on `regions`, `table.regions`
+
                 auto dm_storage = std::dynamic_pointer_cast<StorageDeltaMerge>(storage);
                 auto region_it = table.regions.find(region_id);
-                if (region_it == table.regions.end())
-                    return;
-                HandleRange<HandleID> handle_range = region_it->second.range_in_table;
+                if (dm_storage == nullptr || region_it == table.regions.end())
+                    break;
 
+                HandleRange<HandleID> handle_range = region_it->second.range_in_table;
                 auto dm_handle_range = toDMHandleRange(handle_range);
                 dm_storage->deleteRange(dm_handle_range, context->getSettingsRef());
-                dm_storage->flushCache(*context, dm_handle_range);
+                dm_storage->flushCache(*context, dm_handle_range); // flush to disk
             }
-        }
+        } while (0);
 
         regions.erase(it);
         table.regions.erase(region_id);
         if (table.regions.empty())
         {
+            /// All regions of this table is removed, the storage maybe drop or pd
+            /// move it to another node, we can optimize outdated data.
             table_to_optimize.insert(table_id);
             tables.erase(table_id);
         }
